feat: Update setup script to create learning-portal directory and symlink courses
As specified in the drafted spec for learning materials

The setup script has been modified to create the "learning-portal" directory in the user's local share directory if it doesn't exist. Additionally, a symlink is created for each course directory found in "/usr/share/learning-portal/courses". This ensures that the courses are accessible in the user's local share directory.